pyproject.toml
[tool.poetry]
name = "talk2arxiv-server"
version = "0.1.0"
description = ""
authors = ["Evan Hu <evanhu@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
flask = "^3.0.0"
cohere = "^4.34"
scipdf = {git = "https://github.com/titipata/scipdf_parser"}
pinecone-client = "^2.2.4"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
deptry = "^0.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
